Moved x86 arch folder to i386
