+++ kernel/arch/i386/flush.s
