+++ kernel/arch/i386/linker.ld
