+++ sysroot/usr/include/kernel/arch/i386/regs.hpp
